+++ src/Exception/MissingGeodeticSRIDException.php
@@ -6,12 +6,12 @@ class MissingGeodeticSRIDException extends \RuntimeException
 {
     public function __construct(?int $geometrySrid = null, ?string $message = null)
     {
-        $hint = ' Use the non-geodetic getters and setters instead or make sure to retrieve the geometry with SRID = 4326 or SRID = 0.';
+        $hint = ' Use the non-geodetic getters and setters instead or make sure to retrieve the geometry with a SRID specified in the geodetic_srids config or SRID = 0.';
 
         if ($geometrySrid === null) {
-            $message = $message ?? "The geometry's SRID is not a geodetic SRID.".$hint;
+            $message = $message ?? "The geometry's SRID is not listed in the geodetic_srids config.".$hint;
         } else {
-            $message = $message ?? "SRID {$geometrySrid} is not a geodetic SRID.".$hint;
+            $message = $message ?? "SRID {$geometrySrid} is not listed in the geodetic_srids config.".$hint;
         }
 
         parent::__construct($message);